requirements.txt
APScheduler
aiohttp
async-timeout
attrs
chardet
discord.py
idna
multidict
pip
pytz
setuptools
six
typing-extensions
tzlocal
yarl
